refactor: Move encoder into protocol package
