+++ main/protocol/client/dns-client.go
@@ -1,8 +1,8 @@
 package client
 
 import (
-	"dns-reverse-shell/main/encoder"
 	"dns-reverse-shell/main/protocol"
+	"dns-reverse-shell/main/protocol/encoder"
 	"fmt"
 	"github.com/miekg/dns"
 	"os"
